Added ICallbackEvent, GuiFriends.LMPComparator, improvements to PlayerInfo

+++ src/main/java/latmod/ftbu/core/net/MessageClaimChunk.java
@@ -42,6 +42,6 @@ public class MessageClaimChunk extends MessageLM<MessageClaimChunk>
 		LMPlayerServer p = LMWorldServer.inst.getPlayer(ep);
 		if(m.claim) p.claims.claim(m.dim, m.chunkX, m.chunkZ);
 		else p.claims.unclaim(m.dim, m.chunkX, m.chunkZ, false);
-		return new MessageAreaUpdate(m.chunkX, m.chunkZ, m.dim, (byte)1, p);
+		return new MessageAreaUpdate(m.chunkX, m.chunkZ, m.dim, 1, p);
 	}
 }
